Change server architecture
The server I had was starting sass and babel is the background. This was fine and all, but when I stopped the server, I couldn't also stop those processes easily. So I rewrote the server in ruby to call the shell commands as forked processes and stop them along with the WEBrick server.